backend/requirements.txt
# Core Dependencies - FastAPI
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.8.3

# Text-to-Speech
gTTS==2.5.0

# Audio Processing
pydub==0.25.1

# Environment Variables
python-dotenv==1.0.0

# Additional utilities
aiofiles==23.2.1

# Testing
requests==2.31.0
